feat: add router for registering new users
